Add: some properties
