+++ src/main/resources/application.properties
@@ -1,2 +1,13 @@
-spring.security.user.name = user
-spring.security.user.password = 7229b1b0-788a-4643-9d45-70b8b5ed0b7c
+spring.security.user.name=user
+spring.security.user.password=7229b1b0-788a-4643-9d45-70b8b5ed0b7c
+
+spring.datasource.url=jdbc:postgresql://localhost:5432/words_db
+spring.datasource.username=postgres
+spring.datasource.password=password
+
+spring.flyway.user=postgres
+spring.flyway.password=password
+spring.flyway.enabled=true
+spring.flyway.baseline-on-migrate=true
+spring.flyway.locations=classpath:db/migration
+spring.flyway.schemas=public
