+++ src/main/resources/db/migration/V1__create_word_table.sql
@@ -1,0 +1,4 @@
+CREATE TABLE IF NOT EXISTS word(
+    id BIGINT NOT NULL PRIMARY KEY,
+    name VARCHAR(100) NOT NULL
+);
