--- src/test/java/mvctuto/mvctuto/MvcTutoApplicationTests.java
@@ -1,13 +1,0 @@
-package mvctuto.mvctuto;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class MvcTutoApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
