+++ src/test/java/mvctuto/mvctuto/api/WordsControllerTests.java
@@ -1,0 +1,5 @@
+package mvctuto.mvctuto.api;
+
+
+public class WordsControllerTests {
+}
